sample/producer/requirements.txt
botocore
boto3
StatsD
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-sdk-extension-aws
opentelemetry-instrumentation-botocore
prometheus-client